Address review suggestions: Rename utils.tracing -> utils.contexts
